refactored stuff and now can save transactions after pressing charge

+++ frontend/app/src/main/java/com/example/frontendinternship/domain/model/ProductWithCount.kt
@@ -3,7 +3,7 @@ package com.example.frontendinternship.domain.model
 
 data class ProductWithCount(
     val product: ProductModel = ProductModel(),
-    var count: Int = 1,
+    var count: Double = 1.0,
 ) {
 }
